requirements.txt
torch
torchvision
timm
scikit-learn
pandas
numpy
matplotlib
nibabel
pydicom
opencv-python
Pillow
fastapi
uvicorn[standard]
python-multipart
aiofiles
tqdm
joblib
jinja2
pydantic
scipy
reportlab
pytest
httpx
